grpc and http client call updates
