+++ ai-model/models/nlp/inference/control.py
@@ -30,9 +30,9 @@ if __name__ == '__main__':
     model_name = 'simple'
 
     # There are eight models in the repository directory
-    if len(triton_client.get_model_repository_index().models) != 9:
-        print('FAILED : Repository Index')
-        sys.exit(1)
+    # if len(triton_client.get_model_repository_index().models) != 9:
+    #     print('FAILED : Repository Index')
+    #     sys.exit(1)
     
 
     #NOTE : this only works if calling as explicit 
@@ -40,7 +40,7 @@ if __name__ == '__main__':
     triton_client.load_model(model_name)
     if not triton_client.is_model_ready(model_name):
         print('FAILED : Load Model')
-        sys.exit(1)
+        sys.exit(1) 
 
     triton_client.unload_model(model_name)
     if triton_client.is_model_ready(model_name):
