+++ ai-model/models/nlp/inference/http_client_nli.py
@@ -1,30 +1,31 @@
-from importlib.metadata import version
+import argparse
 import numpy as np
 import sys
-import tritonclient.grpc.model_config_pb2 as mc
-import tritonclient.grpc as tritongrpc
+from functools import partial
+import os
+import tritonclient
+import tritonclient.http as tritonhttp
 from tritonclient.utils import triton_to_np_dtype
 from tritonclient.utils import InferenceServerException
 from transformers import XLMRobertaTokenizer
 from scipy.special import softmax
-from transformers import AutoTokenizer
-import torch
 from datetime import datetime
 
 # load tokeniser
-tokenizer = AutoTokenizer.from_pretrained('xlm-roberta-large')
+R_tokenizer = XLMRobertaTokenizer.from_pretrained('joeddav/xlm-roberta-large-xnli')
 VERBOSE = False
 
-# set up inputs for triton server
+# set up input for triton
 input_name = ['input__0', 'input__1']
 output_name = 'output__0'
 
-def run_inference(premise, model_name='xlm_rob_large_mask', url='127.0.0.1:8001', model_version='1'):
-
+def run_inference(premise,topic, model_name='zst', url='127.0.0.1:8000', model_version='1'):
+    topic = topic
     print(f'\n[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PREMISE: {premise}')
+    print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] TOPIC: {topic}')
 
     # establish connection to triton
-    triton_client = tritongrpc.InferenceServerClient(
+    triton_client = tritonhttp.InferenceServerClient(
         url=url, verbose=VERBOSE)
     print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Connection')
 
@@ -35,53 +36,60 @@ def run_inference(premise, model_name='xlm_rob_large_mask', url='127.0.0.1:8001'
         sys.exit(1) 
     else:
         print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Load Model')
-
-    # get model data
+        
+     # get model data
     model_metadata = triton_client.get_model_metadata(
         model_name=model_name, model_version=model_version)
     model_config = triton_client.get_model_config(
         model_name=model_name, model_version=model_version)    
     print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] CONFIG: {model_config}')
-    
-    # tokenize inputs and format
-    inputs = tokenizer(premise, return_tensors="pt",truncation=True,padding = 'max_length' )
-    input_ids = np.array(inputs['input_ids'], dtype=np.int32)
-    attention_mask = np.array(inputs['attention_mask'], dtype=np.int32)
+
+    # tokenize input
+    input_ids = R_tokenizer.encode(premise, topic, max_length=256, 
+    truncation=True, padding='max_length')
+    print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Tokenize')
+
+    # format inputs
+    input_ids = np.array(input_ids, dtype=np.int32)
+    mask = input_ids != 1
+    mask = np.array(mask, dtype=np.int32)
+    mask = mask.reshape(1, 256) 
+    input_ids = input_ids.reshape(1, 256)
 
     # insert inputs and output(s)
-    print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Tokenize')
-    input0 = tritongrpc.InferInput(input_name[0], (1,  512), 'INT32')
+    input0 = tritonhttp.InferInput(input_name[0], (1,  256), 'INT32')
     input0.set_data_from_numpy(input_ids, binary_data=False)
-    input1 = tritongrpc.InferInput(input_name[1], (1, 512), 'INT32')
-    input1.set_data_from_numpy(attention_mask, binary_data=False)
-    output = tritongrpc.InferRequestedOutput(output_name,  binary_data=False)
+    input1 = tritonhttp.InferInput(input_name[1], (1, 256), 'INT32')
+    input1.set_data_from_numpy(mask, binary_data=False)
+    output = tritonhttp.InferRequestedOutput(output_name,  binary_data=False)
     print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Inputs/Outputs')
 
     # send inputs for inference and recieve output(s)
     response = triton_client.infer(model_name,
-    model_version = model_version, inputs=[input0, input1], outputs=[output])
+    model_version=model_version, inputs=[input0, input1], outputs=[output])
     print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Inference')
 
     # format output(s)
     logits = response.as_numpy('output__0')
-    logits = torch.Tensor(logits)
-    mask_token_index = torch.where(inputs["input_ids"] == tokenizer.mask_token_id)[1]
-    mask_token_logits = logits[0, mask_token_index,:]
-    top_5_tokens = torch.topk(mask_token_logits, 5, dim=1).indices[0].tolist()
-    print(torch.topk(mask_token_logits, 5, dim=1))
-    print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Masks Filled')
+    logits = np.asarray(logits, dtype=np.float32)
+    entail_contradiction_logits = logits[:,[0,2]]
+    probs = softmax(entail_contradiction_logits)
+    true_prob = probs[:,1].item() * 100
+    print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Probability')
 
     # display outputs
-    for token in top_5_tokens:
-        print(f">>>\t {premise.replace(tokenizer.mask_token, tokenizer.decode([token]))}")
-
-    # unload model from client
+    print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] LABEL IS TRUE: {true_prob:0.2f}%')
+    
+    # unload model from triton server
     triton_client.unload_model(model_name)
     if triton_client.is_model_ready(model_name):
         print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] FAILED: Unload Model')
         sys.exit(1)
     else:
-        print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED: Unload Model')
+        print(f'[{(datetime.now()).strftime("%d-%m-%Y %H:%M:%S")}] PASSED : Unload Model')
+    
+
 
 if __name__ == '__main__':
-    run_inference("The dog ran <mask>.",'xlm_rob_large_mask')
+    run_inference("The nebulas dance divinely in our heavenly skies"
+    ,'This text is about space and cosmos','xlm_roberta_zsl')
